히어로 업그레이드 2차

+++ Source/EarthHero/Stat/StatCalculationLibrary.h
@@ -14,7 +14,7 @@ public:
 	static void CalNormalDamage(FStatStructure &HeroStat, float inDamage);
 
 	UFUNCTION()
-	static float CalShooterNormalDamage(FStatStructure &HeroStat, const FHitResult & HitInfo, int32 SH_HeadShot, float SH_NormalDamage);
+	static float CalShooterNormalDamage(FStatStructure &HeroStat, const FHitResult & HitInfo, float SH_HeadShot, float SH_NormalDamage);
 
 	UFUNCTION()
 	static float CalShooterGrenadeDamage(FStatStructure &HeroStat, float SH_GrenadeDamage);
